Tests

+++ ResourceTranslator.CLI/DictionaryFileHelper.cs
@@ -1,4 +1,6 @@
-﻿using Nextended.Core.Helper;
+﻿using Newtonsoft.Json.Linq;
+using Nextended.Core.Helper;
+using ResourceTranslator.CLI.Helper;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -8,7 +10,6 @@ using System.Text.Encodings.Web;
 using System.Text.Json;
 using System.Threading.Tasks;
 using System.Xml;
-using Newtonsoft.Json.Linq;
 using YamlDotNet.Serialization;
 
 namespace ResourceTranslator.CLI
@@ -73,17 +74,20 @@ namespace ResourceTranslator.CLI
 
         private static Task SaveAsJson(IDictionary<string, string> dictionary, string fileName, Encoding encoding)
         {
-            if (File.Exists(fileName))
-                File.Delete(fileName);
-            var dict = JsonDictionaryConverter.ConvertToUnflattenDictionary(dictionary);
+            if (File.Exists(fileName)) File.Delete(fileName);
+
+            var obj = InternalJsonDictionaryConverter.ConvertToUnflattenDictionary(dictionary);
+
             var jsonSerializerOptions = new JsonSerializerOptions
             {
                 Encoder = JavaScriptEncoder.UnsafeRelaxedJsonEscaping,
                 WriteIndented = true
             };
-            
-            string json = JsonSerializer.Serialize(dict, jsonSerializerOptions);
-            File.WriteAllText(fileName, json, encoding);
+
+            File.WriteAllText(fileName, JsonSerializer.Serialize(
+                    JsonDocument.Parse(obj.ToString()).RootElement, jsonSerializerOptions),
+                encoding);
+
             return Task.CompletedTask;
         }
 
@@ -113,9 +117,9 @@ namespace ResourceTranslator.CLI
 
         private static IDictionary<string, string> CreateFromJson(string filename)
         {
-            JObject jsonObject = JObject.Parse(File.ReadAllText(filename));
-            var dict = JsonDictionaryConverter.Flatten(jsonObject);
-            return dict;
+            var json = File.ReadAllText(filename);
+            var obj = JObject.Parse(json);
+            return InternalJsonDictionaryConverter.Flatten(obj);
             //return JsonSerializer.Deserialize<Dictionary<string, string>>(File.ReadAllText(filename));
         }
